port_app/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
cairocffi==0.7.2
CairoSVG==1.0.20
celery==3.1.20
cffi==1.5.2
cssselect==0.9.1
Cython==0.23.4
Django==1.9.2
django-auditlog==0.3.3
django-authtools==1.4.0
django-braces==1.8.1
django-celery==3.1.17
django-jsonfield==0.9.19
django-modeltranslation==0.11
django-widget-tweaks==1.4.1
docutils==0.12
gunicorn==19.4.5
html5lib==0.9999999
jsonfield==1.0.3
kombu==3.0.33
lxml==3.5.0
psycopg2==2.6.1
pycparser==2.14
Pyphen==0.9.4
pytz==2015.7
redis==2.10.5
requests==2.9.1
setproctitle==1.1.9
simplejson==3.8.2
six==1.10.0
suit==2.0.2
tinycss==0.3
WeasyPrint==0.26
xmltodict==0.10.1
